chore: move backend utils and types to common packages
