+++ packages/suite/src/hooks/settings/backends/useCustomBackends.ts
@@ -1,5 +1,5 @@
 import { useSelector } from '@suite-hooks';
-import { getCustomBackends } from '@suite-utils/backend';
+import { getCustomBackends } from '@suite-common/wallet-utils';
 
 export const useCustomBackends = () => {
     const blockchains = useSelector(state => state.wallet.blockchain);
