+++ packages/suite/src/storage/definitions.ts
@@ -2,7 +2,7 @@ import type { DBSchema } from 'idb';
 
 import type { StorageUpdateMessage } from '@trezor/suite-storage';
 import type { State as WalletSettings } from '@wallet-reducers/settingsReducer';
-import type { BackendSettings } from '@wallet-reducers/blockchainReducer';
+import type { BackendSettings } from '@suite-common/wallet-types';
 import type { SuiteState } from '@suite-reducers/suiteReducer';
 import type { AnalyticsState } from '@suite-reducers/analyticsReducer';
 import type { FormState } from '@wallet-types/sendForm';
